+++ file_input.h
@@ -12,10 +12,10 @@ enum {RT_INTERNAL_FIXED_N, RT_INTERNAL_FIXED_MEMORY, RT_WHIZARD};
 typedef struct {
   const char *ref_file;
   int run_type;  
-  long long warmup_trials;
+  size_t warmup_trials;
   int warmup_events;
-  long long internal_events;
-  long long gpu_memory;
+  size_t internal_events;
+  size_t gpu_memory;
   int msq_threads;
   int cb_threads;
   int ab_threads; 
@@ -29,7 +29,7 @@ enum {H_NCHANNELS=0, H_NIN=1, H_NOUT=2, H_NTREES=3, H_NGROVES=4, H_NX=5};
 
 void read_input_json (const char *filename);
 
-long long count_nevents_in_reference_file (const char *ref_file, int n_moment, int filepos);
+size_t count_nevents_in_reference_file (const char *ref_file, int n_moment, int filepos);
 void read_reference_header (const char *ref_file, int *header_data, int *filepos);
 void read_tree_structures (const char *ref_file, int n_trees, int n_prt, int n_prt_out, int n_external, int *filepos);
 void read_reference_momenta (const char *ref_file, int filepos, int n_momenta, int n_x,
